ZBA instructions hooked up and working

+++ core/inst_handlers/b/RvzbaInsts.cpp
@@ -61,10 +61,11 @@ namespace atlas
     {
         const AtlasInstPtr & inst = state->getCurrentInst();
 
-        const uint64_t rs1_val = READ_INT_REG<uint64_t>(state, inst->getRs1());
+        // Ensures that rs1_val only receives the least-significant word of rs1
+        const uint64_t rs1_val = static_cast<uint64_t>(static_cast<uint32_t>(READ_INT_REG<uint64_t>(state, inst->getRs1())));
         const uint64_t rs2_val = READ_INT_REG<uint64_t>(state, inst->getRs2());
 
-        const uint64_t rd_val = ((uint64_t)(uint32_t)(rs1_val)) + rs2_val;
+        const uint64_t rd_val = rs1_val + rs2_val; 
         WRITE_INT_REG<uint64_t>(state, inst->getRd(), rd_val);
 
         return ++action_it;
@@ -87,10 +88,10 @@ namespace atlas
     {
         const AtlasInstPtr & inst = state->getCurrentInst();
 
-        const uint64_t rs1_val = READ_INT_REG<uint64_t>(state, inst->getRs1());
+        const uint64_t rs1_val = static_cast<uint64_t>(static_cast<uint32_t>(READ_INT_REG<uint64_t>(state, inst->getRs1())));
         const uint64_t rs2_val = READ_INT_REG<uint64_t>(state, inst->getRs2());
 
-        const uint64_t rd_val = (rs1_val << 1) + ((uint64_t)(uint32_t)(rs2_val));
+        const uint64_t rd_val = (rs1_val << 1) + rs2_val;
         WRITE_INT_REG<uint64_t>(state, inst->getRd(), rd_val);
 
         return ++action_it;
@@ -113,10 +114,10 @@ namespace atlas
     {
         const AtlasInstPtr & inst = state->getCurrentInst();
 
-        const uint64_t rs1_val = READ_INT_REG<uint64_t>(state, inst->getRs1());
+        const uint64_t rs1_val = static_cast<uint64_t>(static_cast<uint32_t>(READ_INT_REG<uint64_t>(state, inst->getRs1())));
         const uint64_t rs2_val = READ_INT_REG<uint64_t>(state, inst->getRs2());
 
-        const uint64_t rd_val = (rs1_val << 2) + ((uint64_t)(uint32_t)(rs2_val));
+        const uint64_t rd_val = (rs1_val << 2) + rs2_val;
         WRITE_INT_REG<uint64_t>(state, inst->getRd(), rd_val);
 
         return ++action_it;
@@ -139,10 +140,10 @@ namespace atlas
     {
         const AtlasInstPtr & inst = state->getCurrentInst();
 
-        const uint64_t rs1_val = READ_INT_REG<uint64_t>(state, inst->getRs1());
+        const uint64_t rs1_val = static_cast<uint64_t>(static_cast<uint32_t>(READ_INT_REG<uint64_t>(state, inst->getRs1())));
         const uint64_t rs2_val = READ_INT_REG<uint64_t>(state, inst->getRs2());
 
-        const uint64_t rd_val = (rs1_val << 3) + ((uint64_t)(uint32_t)(rs2_val));
+        const uint64_t rd_val = (rs1_val << 3) + rs2_val;
         WRITE_INT_REG<uint64_t>(state, inst->getRd(), rd_val);
 
         return ++action_it;
